feat: 일치하는 로또 번호 개수 계산하는 메소드 리턴타입 변경

+++ src/test/java/lotto/domain/LottoNumbersTest.java
@@ -25,10 +25,10 @@ class LottoNumbersTest {
     @DisplayName("티겟이 당첨번호와 몇개 일치하는지 판별한다.")
     @ParameterizedTest
     @CsvSource(value = {"1,2,3,4,5,6:6","1,2,3,4,5,12:5","1,2,3,4,11,21:4","1,2,3,12,13,14:3", "10,11,12,13,14,15:0"}, delimiter = ':')
-    void count_one_ticket(String input, long expected) {
+    void count_one_ticket(String input, int expected) {
         LottoNumbers winTicket = LottoNumbers.of("1,2,3,4,5,6");
         LottoNumbers buyTicket = LottoNumbers.of(input);
-        long matchCount = winTicket.match(buyTicket);
+        int matchCount = winTicket.match(buyTicket);
 
         assertThat(matchCount).isEqualTo(expected);
     }
